+++ app/javascript/components/great.js
@@ -1,19 +1,16 @@
-import { useDispatch, useSelector } from "react-redux";
-import { fetchUserData } from "../redux/user/userSlice";
-import { useEffect } from "react";
-
-
+import React from "react";
+import { useSelector } from "react-redux";
 
 export const Great = () => {
-    const user = useSelector((state) => state.user.data)    
-    const dispatch = useDispatch;
-    useEffect(() => {
-        dispatch(fetchUserData());
-    }, [dispatch]);
-    //const data = useSelector((state) => state.user.data)
-    return (
-        <h1>{user.greeting}</h1>
-    );
+  const user = useSelector((state) => state.user)
+  let valeur = "";
+
+  if (user.status == "succeeded") {
+    valeur = user.data.greeting
+  }
+  return(
+    <h1>{valeur}</h1>
+  );
 }
 
 export default Great;
