+++ app/javascript/redux/store.js
@@ -1,5 +1,6 @@
+import React from "react";
 import { configureStore } from "@reduxjs/toolkit";
-import { userReducer } from "./user/userSlice"
+import  userReducer  from "./user/userSlice"
 
 export const store = configureStore({
     reducer: {
