+++ app/javascript/redux/user/userSlice.js
@@ -1,18 +1,14 @@
+import React from "react";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import axios from "axios";
 
 const GREETING_URL = 'http://127.0.0.1:3000/greeting';
 
 
-export const fetchUserData = createAsyncThunk("user/fetchUserData", () => {
-    return axios
-      .get(GREETING_URL)
-      .then((response) => {
-       console.log(response) 
-      })
-  }
-);
-
+  export const fetchUserData = createAsyncThunk("user/fetchUserData", async () => {
+    const response = await axios.get(GREETING_URL);
+    return response.data
+  });
   const userSlice = createSlice({
     name: "user",
     initialState: {
